+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+
+ - Fish,AOT: Implemented fallbacks
+
 ### Fixed
 
  - Fish,JIT: Descriptions were wrongly being made part of the completion